css/css-align/parsing/place-items-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: place-items sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-place-items">
<meta name="assert" content="place-items supports the full grammar '<align-items> <justify-items>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
'use strict';

function test_place_content_value(shorthandValue, longhandValues) {
  const shorthandName = 'place-items';
  const longhandNames = [
      'align-items',
      'justify-items'
  ];
  test_shorthand_value(shorthandName, shorthandValue, longhandNames, longhandValues);
}

test_place_content_value('normal', [
    'normal',
    'normal'
]);

test_place_content_value('first baseline', [
    'baseline',
    'baseline'
]);

test_place_content_value('last baseline flex-start', [
    'last baseline',
    'flex-start'
]);

test_place_content_value('stretch right legacy', [
    'stretch',
    'legacy right'
]);
</script>
</body>
</html>
